+++ resolwe/flow/signals.py
@@ -10,11 +10,12 @@ from asgiref.sync import async_to_sync
 from django.conf import settings
 from django.db import transaction
 from django.db.models import Count
-from django.db.models.signals import post_save, pre_delete
+from django.db.models.signals import post_delete, post_save, pre_delete
 from django.dispatch import receiver
 
 from resolwe.flow.managers import manager
-from resolwe.flow.models import Data, Entity
+from resolwe.flow.models import Data, Entity, Relation
+from resolwe.flow.models.entity import RelationPartition
 
 
 def commit_signal(data_id):
@@ -39,3 +40,22 @@ def delete_entity(sender, instance, **kwargs):
     """Delete Entity when last Data object is deleted."""
     # 1 means that the last Data object is going to be deleted.
     Entity.objects.annotate(num_data=Count('data')).filter(data=instance, num_data=1).delete()
+
+
+# NOTE: m2m_changed signal cannot be used because of a bug:
+# https://code.djangoproject.com/ticket/17688
+@receiver(post_delete, sender=RelationPartition)
+def delete_relation(sender, instance, **kwargs):
+    """Delete the Relation object when the last Entity is removed."""
+    def process_signal(relation_id):
+        """Get the relation and delete it if it has no entities left."""
+        try:
+            relation = Relation.objects.get(pk=relation_id)
+        except Relation.DoesNotExist:
+            return
+
+        if relation.entities.count() == 0:
+            relation.delete()
+
+    # Wait for partitions to be recreated.
+    transaction.on_commit(lambda: process_signal(instance.relation_id))