+++ docs/CHANGELOG.rst
@@ -23,6 +23,7 @@ Changed
   - requireing uniqe combination of ``collection`` and ``category``
   - renaming partition's ``position`` to ``label``
   - adding (integer) ``position`` to partition (used for sorting)
+  - deleting ``Relation`` when the last ``Entity`` is removed
 - **BACKWARD INCOMPATIBLE:** Remove rarely used parameters of the ``register``
   command ``--path`` and ``--schemas``.
 - Omit ``current_user_permissions`` field in serialization if only a subset of